requirements.txt
alembic==1.8.0
bcrypt==3.2.2
cffi==1.15.1
click==8.1.3
decorator==5.1.1
dnspython==2.2.1
dominate==2.6.0
email-validator==1.2.1
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.1
Flask-Migrate==2.6.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.8.0
infinity==1.5
intervals==0.9.2
itsdangerous==1.1.0
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
pycparser==2.21
PyJWT==2.4.0
six==1.16.0
SQLAlchemy==1.4.39
validators==0.20.0
visitor==0.1.3
Werkzeug==2.1.2
WTForms==2.3.3
WTForms-Components==0.10.5
zipp==3.8.0
